requirements.txt
ultralytics==8.0.0
opencv-python==4.7.0.72
openai==0.27.8
pyttsx3==2.90
python-dotenv==1.0.0
pillow==9.5.0
numpy==1.24.3
llama-cpp-python==0.2.23